+++ gulpfile.js
@@ -63,7 +63,7 @@ gulp.task('lint', function () {
     // So, it's best to have gulp ignore the directory as well.
     // Also, Be sure to return the stream from the task;
     // Otherwise, the task may end before the stream has finished.
-    return gulp.src(['./widgets/**/*.js*','WRIO-InternetOS/**/*.js'])
+    return gulp.src(['src/**/*.js'])
         // eslint() attaches the lint output to the "eslint" property
         // of the file object so it can be used by other modules.
         .pipe(eslint())
@@ -87,7 +87,7 @@ var version = getVersion();
 gulp.task('babel-client', function() {
 
     var preloader = browserify({
-        entries: './WRIO-InternetOS/preloader.js',
+        entries: './src/preloader.js',
         debug: true
     })
       .transform(babelify)
@@ -114,7 +114,7 @@ gulp.task('babel-client', function() {
 
 
     var main = browserify({
-            entries: './WRIO-InternetOS/main.js',
+            entries: './src/main.js',
             debug: true
         })
         .transform(babelify)
@@ -169,14 +169,13 @@ gulp.task('default', ['lint','babel-client']);
 
 gulp.task('watch', ['default'], function() {
     gulp.watch([
-        'WRIO-InternetOS/**/*.*',
-        'widgets/**/*.*'
+        'src/**/*.*',
     ], ['babel-client']);
 });
 
 gulp.task('watchDev', ['default'], function() {
     gulp.watch([
-        'WRIO-InternetOS/**/*.*','widgets/**/*.*'
+        'src/**/*.*',
     ], ['babel-client']);
 });
