+++ src/core/actions/WindowActions.js
