+++ src/core/addBootstrapLink.js
