+++ src/core/components/CenterContent.js
