+++ src/core/components/CreateCover.js
