+++ src/core/components/CreateDomCenter.js
@@ -1,23 +1,23 @@
 import {getServiceUrl,getDomain} from '../servicelocator.js';
 import React from 'react';
 import Reflux from 'reflux';
-import Login from '../../../widgets/Login.jsx';
-import Chess from '../../../widgets/Chess.jsx';
-import Core from '../../../widgets/Core.jsx';
-import Details from '../../../widgets/Details.jsx';
+import Login from '../../widgets/Login.jsx';
+import Chess from '../../widgets/Chess.jsx';
+import Core from '../../widgets/Core.jsx';
+import Details from '../../widgets/Details.jsx';
 import {importUrl} from '../global';
 import {theme} from '../global';
 import CreateBreadcrumb from './CreateBreadcrumb';
-import CreateTitter from '../../../widgets/titter.jsx';
+import CreateTitter from '../../widgets/titter.jsx';
 import CenterContent from './CenterContent';
 import StoreLd from '../store/center';
 import classNames from 'classnames';
-import ActionMenu from '../../../widgets/Plus/actions/menu';
-import StoreMenu from '../../../widgets/Plus/stores/menu';
+import ActionMenu from '../../widgets/Plus/actions/menu';
+import StoreMenu from '../../widgets/Plus/stores/menu';
 import UrlMixin from '../mixins/UrlMixin';
-import CreateTransactions from '../../../widgets/transactions.jsx';
+import CreateTransactions from '../../widgets/transactions.jsx';
 import CenterActions from '../actions/center';
-import PlusStore from '../../../widgets/Plus/stores/PlusStore.js';
+import PlusStore from '../../widgets/Plus/stores/PlusStore.js';
 import WindowActions from '../actions/WindowActions.js';
 import {AlertWelcome, AlertWarning} from './Alerts.js';
 import UserStore from '../store/UserStore.js';
