+++ src/core/components/CreateDomRight.js
@@ -2,10 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import UrlMixin from '../mixins/UrlMixin';
 import classNames from 'classnames';
-import ActionMenu from '../../../widgets/Plus/actions/menu';
+import ActionMenu from '../../widgets/Plus/actions/menu';
 import CreateInfoTicket from './CreateInfoTicket';
 import CreateControlButtons from './CreateControlButtons';
-import StoreMenu from '../../../widgets/Plus/stores/menu';
+import StoreMenu from '../../widgets/Plus/stores/menu';
 import Reflux from 'reflux';
 import _ from 'lodash';
 import center from '../actions/center';
