+++ src/core/global.js
