+++ src/core/jsonld/scripts.js
