+++ src/core/mixins/mentions.js
