+++ src/core/servicelocator.js
