+++ src/core/store/CrossStorageFactory.js
