+++ src/core/store/WindowMessage.js
