+++ src/widgets/Plus/Plus.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Reflux from 'reflux';
 import PlusStore from './stores/PlusStore.js';
-import UserStore from '../../WRIO-InternetOS/js/store/UserStore.js';
+import UserStore from '../../core/store/UserStore.js';
 import actions from './actions/jsonld';
 import ActionMenu from './actions/menu';
 import StoreMenu from './stores/menu';
