+++ src/widgets/Plus/actions/menu.js
