+++ src/widgets/Plus/stores/menu.js
